+++ DESCRIPTION
@@ -26,7 +26,7 @@ Description: Bland (2009) <doi:10.1136/bmj.b3985> recommended to
     precision based sample size calculations. For a given sample size, the 
     functions will return the precision (width of the confidence interval), and 
     vice versa. 
-License: GPL-3
+License: GPL-3 + file LICENSE
 URL: https://github.com/CTU-Bern/presize
 BugReports: https://github.com/CTU-Bern/presize/issues
 Encoding: UTF-8
