### 2021-02-11 #3030 @aghaynes
Thanks for the feedback @arfon. I've added those things. I also noticed that the version on github is actually 0.2.0 rather than 0.1.3 which is on CRAN.

On a different point, Whedon seems to think that I've only made 2 commits... that is not even remotely correct... I've committed many more times...
