### 2021-02-20 #3030 @xuanxu
In the paper metadata the `orcid` field for Odile Stalder is empty. There are two options to fix the error, adding a value or removing the orcid field for that author.
