### 2021-03-04 #3030 @majensen
@seandavi - would you be willing to review [this work](#3030) for [JOSS](https://joss.theoj.org)? If not, would you be able to recommend an expert from your group or elsewhere? I really appreciate  your consideration - @majensen
