### 2021-03-08 #3030 @majensen
Hey @amoeba, do you have time and inclination to review [this work](#3030) for [JOSS](https://joss.theoj.org)? Would be awesome.
