### 2021-03-11 87e38de
defaults for sliders
