+++ inst/shinyApp/ui/18_lrneg.R
@@ -8,11 +8,12 @@ lrnpage <- tabItem(tabName = "lrneg",
         tags$br(),
         "",
         h4("Please enter the following"),
-        sliderInput("lrn_prev", "Prevalence", min = 0, max = 1, value = NULL),
+        sliderInput("lrn_prev", "Prevalence",
+                    min = 0, max = 1, value = .5),
         sliderInput("lrn_sens", "Sensitivity",
-                    min = 0, max = 1, value = NULL),
+                    min = 0, max = 1, value = .5),
         sliderInput("lrn_spec", "Specificity",
-                    min = 0, max = 1, value = NULL),
+                    min = 0, max = 1, value = .5),
         h4("Please enter one of the following"),
         numericInput("lrn_n", "Total sample size",
                      value = NULL),
