+++ inst/shinyApp/ui/05_riskdiff.R
@@ -5,9 +5,9 @@ riskdiffpage <- tabItem(tabName = "riskdiff",
         "Enter the proportions of events you expect in the groups. If you intend to use uneven allocation ratios (e.g. 2 allocated to group 1 for each participant allocated to group 2), adjust the allocation ratio accordingly. To estimate the confidence interval width from a number of events, enter the number of events in 'Number of events'. To estimate the number of observations required to get a confidence interval width of X, enter the width in 'Confidence interval width'.",
         h4("Please enter the following"),
         sliderInput("riskdiff_p1", "Proportion of events in group 1",
-                    value = NULL, min = 0, max = 1),
+                    value = .5, min = 0, max = 1),
         sliderInput("riskdiff_p2", "Proportion of events in group 2",
-                    value = NULL, min = 0, max = 1),
+                    value = .4, min = 0, max = 1),
         numericInput("riskdiff_r", "Allocation ratio", value = 1),
         "N1 / N2",
         h4("Please enter one of the following"),
