+++ inst/shinyApp/ui/17_lrpos.R
@@ -8,11 +8,12 @@ lrppage <- tabItem(tabName = "lrpos",
         tags$br(),
         "",
         h4("Please enter the following"),
-        sliderInput("lrp_prev", "Prevalence", min = 0, max = 1, value = NULL),
+        sliderInput("lrp_prev", "Prevalence",
+                    min = 0, max = 1, value = .5),
         sliderInput("lrp_sens", "Sensitivity",
-                    min = 0, max = 1, value = NULL),
+                    min = 0, max = 1, value = .5),
         sliderInput("lrp_spec", "Specificity",
-                    min = 0, max = 1, value = NULL),
+                    min = 0, max = 1, value = .5),
         h4("Please enter one of the following"),
         numericInput("lrp_n", "Total sample size",
                      value = NULL),
