+++ inst/shinyApp/ui/13_sensitivity.R
@@ -5,8 +5,10 @@ senspage <- tabItem(tabName = "sens",
         "Sensitivity is the proportion of positive test results that are identified as such. It is also known as the true positive rate, recall or probability of detection. It is actually a simple proportion, but as the total sample size, rather than the number of cases, is typically of interect this function requires an estimate of the prevalence of cases.",
         tags$br(),
         h4("Please enter the following"),
-        sliderInput("sens_sens", "Sensitivity", min = 0, max = 1, value = NULL),
-        sliderInput("sens_prev", "Prevalence", min = 0, max = 1, value = NULL),
+        sliderInput("sens_sens", "Sensitivity",
+                    min = 0, max = 1, value = .5),
+        sliderInput("sens_prev", "Prevalence",
+                    min = 0, max = 1, value = .4),
         tags$br(),
         h4("Please enter one of the following"),
         numericInput("sens_ntot", "Total sample size",
