+++ inst/shinyApp/ui/07_riskratio.R
@@ -6,9 +6,9 @@ riskrpage <- tabItem(tabName = "riskratio",
         tags$br(),
         h4("Please enter the following"),
         sliderInput("riskratio_p1", "Proportion of events in group 1",
-                    value = NULL, min = 0, max = 1),
+                    value = .5, min = 0, max = 1),
         sliderInput("riskratio_p2", "Proportion of events in group 2",
-                    value = NULL, min = 0, max = 1),
+                    value = .4, min = 0, max = 1),
         numericInput("riskratio_r", "Allocation ratio", value = 1),
         "(N2 / N1)",
         h4("Please enter one of the following"),
