+++ inst/shinyApp/ui/15_auc.R
@@ -6,8 +6,10 @@ aucpage <- tabItem(tabName = "auc",
         plotOutput("auc_fig"),
         tags$br(),
         "Please enter the following",
-        sliderInput("auc_auc", "AUC", min = 0, max = 1, value = NULL),
-        sliderInput("auc_prev", "Prevalence", min = 0, max = 1, value = NULL),
+        sliderInput("auc_auc", "AUC",
+                    min = 0, max = 1, value = .5),
+        sliderInput("auc_prev", "Prevalence",
+                    min = 0, max = 1, value = .4),
         tags$br(),
         "Please enter one of the following",
         numericInput("auc_n", "Total sample size",
