+++ inst/shinyApp/ui/06_oddsratio.R
@@ -6,9 +6,9 @@ orpage <- tabItem(tabName = "or",
         tags$br(),
         h4("Please enter the following"),
         sliderInput("or_p1", "Proportion of events in group 1*",
-                    value = NULL, min = 0, max = 1),
+                    value = .5, min = 0, max = 1),
         sliderInput("or_p2", "Proportion of events in group 2*",
-                    value = NULL, min = 0, max = 1),
+                    value = .4, min = 0, max = 1),
         numericInput("or_r", "Allocation ratio*", value = 1),
         "(N2 / N1)",
         h4("Please enter one of the following"),
