### 2021-03-11 ac2c149
resettable content ui files 1 to 7
