+++ inst/shinyApp/ui/07_riskratio.R
@@ -12,10 +12,9 @@ riskrpage <- tabItem(tabName = "riskratio",
         numericInput("riskratio_r", "Allocation ratio", value = 1),
         "(N2 / N1)",
         h4("Please enter one of the following"),
-        numericInput("riskratio_n1", "Number of observations in group 1", value = NULL),
-        "(the number of observations in group 2 is estimated based on group 1 and the allocation ratio)",
-        numericInput("riskratio_ciwidth", "Confidence interval width",
-                     value = NULL),
+        uiOutput("riskratio_resetable_input"),
+        actionButton("riskratio_reset_input",
+                     "Reset 'Number of observations' or 'Confidence interval width'"),
         h4("Other settings"),
         selectInput("riskratio_method", label = "Method",
                     choices = c("Koopman" = "koopman",
