+++ inst/shinyApp/ui/03_rate.R
@@ -10,9 +10,12 @@ ratepage <- tabItem(
       "rate_r", "Rate", value = .5, min = 0, max = 1
     ),
     h4("Please enter one of the following"),
-    numericInput("rate_x", "Number of Events", value = NULL),
-    numericInput("rate_ciwidth", "Confidence interval width",
-                 value = NULL),
+    uiOutput("rate_resetable_input"),
+    actionButton("rate_reset_input",
+                 "Reset 'Number of events' or 'Confidence interval width'"),
+    # numericInput("rate_x", "Number of Events", value = NULL),
+    # numericInput("rate_ciwidth", "Confidence interval width",
+    #              value = NULL),
     h4("Other settings"),
     selectInput("rate_method",
       label = "Confidence interval method",
