+++ inst/shinyApp/ui/05_riskdiff.R
@@ -11,10 +11,13 @@ riskdiffpage <- tabItem(tabName = "riskdiff",
         numericInput("riskdiff_r", "Allocation ratio", value = 1),
         "N1 / N2",
         h4("Please enter one of the following"),
-        numericInput("riskdiff_n1", "Number of observations in group 1", value = NULL),
-        "(the number of observations in group 2 is estimated based on group 1 and the allocation ratio)",
-        numericInput("riskdiff_ciwidth", "Confidence interval width",
-                     value = NULL),
+        # numericInput("riskdiff_n1", "Number of observations in group 1", value = NULL),
+        # "(the number of observations in group 2 is estimated based on group 1 and the allocation ratio)",
+        # numericInput("riskdiff_ciwidth", "Confidence interval width",
+        #              value = NULL),
+        uiOutput("riskdiff_resetable_input"),
+        actionButton("riskdiff_reset_input",
+                     "Reset 'Number of observations in group 1' or 'Confidence interval width'"),
         h4("Other settings"),
         selectInput("riskdiff_method", label = "Method",
                     choices = c("Newcombe" = "newcombe",
