+++ inst/shinyApp/ui/02_proportion.R
@@ -9,10 +9,13 @@ proportionpage <- tabItem(tabName = "proportion",
                     value = .5, min = 0, max = 1),
         tags$br(),
         h4("Please enter one of the following"),
-        numericInput("prop_n", "Number of observations",
-                     value = NULL),
-        numericInput("prop_ciwidth", "Confidence interval width",
-                     value = NULL, min = 0, max = 1),
+        # numericInput("prop_n", "Number of observations",
+        #              value = NULL),
+        # numericInput("prop_ciwidth", "Confidence interval width",
+        #              value = NULL, min = 0, max = 1),
+        uiOutput("prop_resetable_input"),
+        actionButton("prop_reset_input",
+                     "Reset 'Number of observations' or 'Confidence interval width'"),
         h4("Other settings"),
         "The Wilson confidence interval is recommended, but others are available.",
         selectInput("prop_method", label = "Confidence interval method",
