+++ inst/shinyApp/ui/06_oddsratio.R
@@ -12,10 +12,13 @@ orpage <- tabItem(tabName = "or",
         numericInput("or_r", "Allocation ratio*", value = 1),
         "(N2 / N1)",
         h4("Please enter one of the following"),
-        numericInput("or_n1", "Number of observations in group 1", value = NULL),
-        "(the number of observations in group 2 is estimated based on group 1 and the allocation ratio)",
-        numericInput("or_ciwidth", "Confidence interval width",
-                     value = NULL),
+        uiOutput("or_resetable_input"),
+        actionButton("or_reset_input",
+                     "Reset 'Number of observations' or 'Confidence interval width'"),
+        # numericInput("or_n1", "Number of observations in group 1", value = NULL),
+        # "(the number of observations in group 2 is estimated based on group 1 and the allocation ratio)",
+        # numericInput("or_ciwidth", "Confidence interval width",
+        #              value = NULL),
         h4("Other settings"),
         selectInput("or_method", label = "Method",
                     choices = c("Gart" = "gart",
