+++ inst/shinyApp/ui/14_specificity.R
@@ -9,10 +9,9 @@ specpage <- tabItem(tabName = "spec",
         sliderInput("spec_prev", "Prevalence",
                     min = 0, max = 1, value = .4),
         h4("Please enter one of the following"),
-        numericInput("spec_ntot", "Total sample size",
-                     value = NULL),
-        numericInput("spec_ciwidth", "Confidence interval width",
-                     value = NULL, min = 0, max = 1),
+        uiOutput("spec_resetable_input"),
+        actionButton("spec_reset_input",
+                     "Reset 'Total sample size' or 'Confidence interval width'"),
         h4("Other settings"),
         selectInput("spec_method", label = "Confidence interval method",
                     choices = c("Wilson" = "wilson",
