+++ inst/shinyApp/ui/10_icc.R
@@ -8,9 +8,9 @@ iccpage <- tabItem(tabName = "icc",
                      value = 0, min = 0, max = 1),
         numericInput("icc_k", "Number of observations per subject", value = NULL),
         h4("Please enter one of the following"),
-        numericInput("icc_n", "Number of subjects", value = NULL),
-        numericInput("icc_ciwidth", "Confidence interval width",
-                     value = NULL, min = 0, max = 1),
+        uiOutput("icc_resetable_input"),
+        actionButton("icc_reset_input",
+                     "Reset 'Number of subjects' or 'Confidence interval width'"),
         h4("Results"),
         verbatimTextOutput("icc_out"),
         tableOutput("icc_tab"),
