+++ inst/shinyApp/ui/17_lrpos.R
@@ -15,11 +15,9 @@ lrppage <- tabItem(tabName = "lrpos",
         sliderInput("lrp_spec", "Specificity",
                     min = 0, max = 1, value = .5),
         h4("Please enter one of the following"),
-        numericInput("lrp_n", "Total sample size",
-                     value = NULL),
-        "The number in each group (e.g. diseased and healthy) is calculated as sample size * prevalence (group 1, which might be diseased people) and sample size * 1-prevalence (group 2).",
-        numericInput("lrp_ciwidth", "Confidence interval width",
-                     value = NULL, min = 0, max = 1),
+        uiOutput("lrp_resetable_input"),
+        actionButton("lpr_reset_input",
+                     "Reset 'Total sample size' or 'Confidence interval width'"),
 
         h4("Results"),
         verbatimTextOutput("lrp_out"),
