+++ inst/shinyApp/ui/08_rateratio.R
@@ -12,11 +12,11 @@ raterpage <- tabItem(tabName = "rateratio",
         numericInput("rateratio_r", "Allocation ratio", value = 1, step = .2),
         "(N2 / N1)",
         h4("Please enter one of the following"),
+        uiOutput("rateratio_resetable_input"),
+        actionButton("rateratio_reset_input",
+                     "Reset 'Number of observations' or 'Confidence interval width'"),
 
-        numericInput("rateratio_n_exp", "Number of observations in the exposed group", value = NULL, step = .1),
-        "(the number of observations in group 2 is estimated based on the number of exposed individuals and the allocation ratio)",
-        numericInput("rateratio_ciwidth", "Upper-lower ratio",
-                     value = NULL, step = .1),
+
         h4("Results"),
         verbatimTextOutput("rateratio_out"),
         tableOutput("rateratio_tab"),
