+++ inst/shinyApp/ui/18_lrneg.R
@@ -15,11 +15,9 @@ lrnpage <- tabItem(tabName = "lrneg",
         sliderInput("lrn_spec", "Specificity",
                     min = 0, max = 1, value = .5),
         h4("Please enter one of the following"),
-        numericInput("lrn_n", "Total sample size",
-                     value = NULL),
-        "The number in each group (e.g. diseased and healthy) is calculated as sample size * prevalence (group 1, which might be diseased people) and sample size * 1-prevalence (group 2).",
-        numericInput("lrn_ciwidth", "Confidence interval width",
-                     value = NULL, min = 0, max = 1),
+        uiOutput("lrn_resetable_input"),
+        actionButton("lpn_reset_input",
+                     "Reset 'Total sample size' or 'Confidence interval width'"),
 
         h4("Results"),
         verbatimTextOutput("lrn_out"),
