+++ inst/shinyApp/ui/13_sensitivity.R
@@ -11,10 +11,9 @@ senspage <- tabItem(tabName = "sens",
                     min = 0, max = 1, value = .4),
         tags$br(),
         h4("Please enter one of the following"),
-        numericInput("sens_ntot", "Total sample size",
-                     value = NULL),
-        numericInput("sens_ciwidth", "Confidence interval width",
-                     value = NULL, min = 0, max = 1),
+        uiOutput("sens_resetable_input"),
+        actionButton("sens_reset_input",
+                     "Reset 'Total sample size' or 'Confidence interval width'"),
         h4("Optional parameters"),
         selectInput("sens_method", label = "Confidence interval method",
                     choices = c("Wilson" = "wilson",
