+++ inst/shinyApp/app.R
@@ -255,6 +255,14 @@ server <- function(input, output, session) {
         tmp1$column[tmp1$column == "ar"] <- "r"
         tmp1[na.omit(match(names(tmp), tmp1$column)),]
     })
+    output$rateratio_resetable_input <- renderUI({
+        times <- input$rateratio_reset_input
+        div(id=letters[(times %% length(letters)) + 1],
+            numericInput("rateratio_n_exp", "Number of observations in the exposed group", value = NULL, step = .1),
+            "(the number of observations in group 2 is estimated based on the number of exposed individuals and the allocation ratio)",
+            numericInput("rateratio_ciwidth", "Upper-lower ratio",
+                         value = NULL, step = .1))
+    })
 
     # correlation coefficient ----
     output$cor_out <- renderPrint(cor_fn(input))
@@ -266,6 +274,14 @@ server <- function(input, output, session) {
         tmp1$column[tmp1$column == "cr"] <- "r"
         tmp1[na.omit(match(names(tmp), tmp1$column)),]
     })
+    output$cor_resetable_input <- renderUI({
+        times <- input$cor_reset_input
+        div(id=letters[(times %% length(letters)) + 1],
+            numericInput("cor_n", "Number of observations", value = NULL),
+            numericInput("cor_ciwidth", "Confidence interval width",
+                         value = NULL, min = 0, max = 1)
+            )
+    })
 
     # ICC ----
     output$icc_out <- renderPrint(icc_fn(input))
@@ -274,6 +290,14 @@ server <- function(input, output, session) {
         tmp <- icc_fn(input)
         tmp1 <- res_vars[res_vars$column %in% names(tmp),]
         tmp1[na.omit(match(names(tmp), tmp1$column)),]
+    })
+    output$icc_resetable_input <- renderUI({
+        times <- input$icc_reset_input
+        div(id=letters[(times %% length(letters)) + 1],
+            numericInput("icc_n", "Number of subjects", value = NULL),
+            numericInput("icc_ciwidth", "Confidence interval width",
+                         value = NULL, min = 0, max = 1)
+        )
     })
 
     # limit of agreement ----
@@ -308,6 +332,15 @@ server <- function(input, output, session) {
         tmp1 <- res_vars[res_vars$column %in% names(tmp),]
         tmp1[na.omit(match(names(tmp), tmp1$column)),]
     })
+    output$limit_resetable_input <- renderUI({
+        times <- input$limit_reset_input
+        div(id=letters[(times %% length(letters)) + 1],
+            numericInput("limit_n", "Sample size",
+                         value = NULL),
+            numericInput("limit_ciwidth", "Confidence interval width",
+                         value = NULL, min = 0, max = 1)
+        )
+    })
 
     # kappa ----
     output$kappa_out <- renderPrint(kappa_fn(input))
@@ -316,6 +349,13 @@ server <- function(input, output, session) {
         tmp <- kappa_fn(input)
         tmp1 <- res_vars[res_vars$column %in% names(tmp),]
         tmp1[na.omit(match(names(tmp), tmp1$column)),]
+    })
+    output$kappa_resetable_input <- renderUI({
+        times <- input$kappa_reset_input
+        div(id=letters[(times %% length(letters)) + 1],
+            numericInput("kappa_n", "Sample size", min = 0, value = NULL),
+            numericInput("kappa_ciwidth", "Confidence interval width", min = 0, value = NULL)
+        )
     })
 
     # sens ----
@@ -328,6 +368,16 @@ server <- function(input, output, session) {
         tmp1$column[tmp1$column == "nsens"] <- "n"
         tmp1[na.omit(match(names(tmp), tmp1$column)),]
     })
+    output$sens_resetable_input <- renderUI({
[…]
