+++ inst/shinyApp/ui/15_auc.R
@@ -12,10 +12,9 @@ aucpage <- tabItem(tabName = "auc",
                     min = 0, max = 1, value = .4),
         tags$br(),
         "Please enter one of the following",
-        numericInput("auc_n", "Total sample size",
-                     value = NULL),
-        numericInput("auc_ciwidth", "Confidence interval width",
-                     value = NULL, min = 0, max = 1),
+        uiOutput("auc_resetable_input"),
+        actionButton("auc_reset_input",
+                     "Reset 'Total sample size' or 'Confidence interval width'"),
         tags$hr(),
         verbatimTextOutput("auc_out"),
         tableOutput("auc_tab"),
