+++ .github/workflows/R-CMD-release.yaml
@@ -1,6 +1,6 @@
-on: [push, pull_request]
+on: [push, pull_request, workflow_dispatch]
 
-name: R-CMD-check
+name: R-CMD-release
 
 jobs:
   R-CMD-check:
