### 2021-03-11 7a989bc
swap platform logo to png
