+++ README.md
@@ -4,7 +4,7 @@ presize
 =======
 
 [![](https://www.r-pkg.org/badges/version/presize?color=green)](https://cran.r-project.org/package=presize)
-[![](https://img.shields.io/badge/dev%20version-0.2.0-blue.svg)](https://github.com/CTU-Bern/presize)
+[![](https://img.shields.io/badge/dev%20version-0.2.1-blue.svg)](https://github.com/CTU-Bern/presize)
 ![travis](https://travis-ci.com/CTU-Bern/presize.svg?branch=master)
 [![AppVeyor Build
 Status](https://ci.appveyor.com/api/projects/status/github/CTU-Bern/presize?branch=master&svg=true)](https://ci.appveyor.com/project/CTU-Bern/presize)
@@ -105,6 +105,6 @@ Funding
 `presize` was largely developed at CTU Bern, with collaboration from CTU
 Basel. Funding was provided by the Swiss Clinical Trial Organisation.
 
-<img src="inst/fig/SCTO_Platform_Logo_2020_RZ_SM.jpg" width="148" />
+<img src="inst/fig/SCTO_Platform_Logo_2020_RZ_SM.png" width="148" />
 
 <!-- ![](man/fig/scto_ctu_member_cmyk.jpg) -->
