### 2021-03-12 #3030 @majensen
@PatrickRWright - any chance you would be willing to review [this work](#3030) for [JOSS](https://joss.theoj.org)? Please have a look at [these reviewing guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) for some additional info. Thanks for considering-
