### 2021-03-16 #3030 @majensen
@TomKellyGenetics - would you have time to review [this work](#3030) for [JOSS](https:://joss.theoj.org)? Much appreciated! --@majensen
