### 2021-03-17 #3118 @whedon
[REVIEW]: `presize`: An R-package for precision-based sample size calculation in clinical research
**Submitting author:** @aghaynes (<a href="http://orcid.org/0000-0003-1374-081X">Alan G Haynes</a>)
**Repository:** <a href="https://github.com/CTU-Bern/presize" target ="_blank">https://github.com/CTU-Bern/presize</a>
**Version:** 0.2.3
