### 2021-03-22 CTU-Bern/presize#61 @amoeba
JOSS review
**Submission:** https://github.com/openjournals/joss-reviews/issues/3118
**Status:** Conditional accept

Hey @aghaynes & @majensen,

I've finished my review and was able to check off most items but a few remain.
The submission is in very good shape overall and I think the changes are minor.

First off, kudos @aghaynes to you and your team for producing such a nicely put together package.
Your description of the background for this work and your accompanying documentation made it easy for me to get familiar with the technique and explore the functionality of the package.
The documentation looked very complete and the Shiny app is nice touch, especially in the way it automatically generates R code to match.

Below are the items I'd like to see addressed before I can change my review to an Accept:

## Contribution and authorship

I see different sets of authors in a few places and wanted to make sure everything was complete and correct:

- Paper: Haynes, Lenz, Stalder, Limacher
- R package: Lenz [aut], Haynes [cre,aut], Limacher [aut], Stalder [ctb]
- git commit history: Haynes, Lenz, Stalder (by order of commits)
- pkgdown site: Lenz, Haynes, Limacher

First, could you clarify Limacher's association and contribution to the package? Second, is the above information (including ordering) correct?

## Functionality

Everything worked great except the Shiny app.
When I install the package, load the package, and attempt to run the Shiny app, I get an error and the app does not launch:

> Error: In sliderInput(): `min`, `max`, and `value` cannot be NULL, NA, or empty. 24.
> stop(call. = FALSE, sprintf("In %s(): `min`, `max`, and `value` cannot be NULL, NA, or empty.", fun)) 23.

I had a hunch and checked to see if the Shiny app worked under the development version and it does.
I assume the authors are aware of this but I think it warrants ensuring JOSS archives a fixed version of the software and not 0.1.3.

Let me know if you can reproduce my issue. If so, I think generating a new release and updating this submission is warranted.

## Community guidelines

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I wasn't able to find any of these three bits of information in the README file.
Please add them.

## Paper

The paper is overall well-written and quite useful.
However, I believe it doesn't fit in with the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) in two ways:

1. It does not contain an initial, high-level summary paragraph entitled Summary
2. It contains too much detail about the package. That said, the information that falls outside the scope of the JOSS paper is very useful information and could be moved elsewhere

## Other, non-review details

These are outside the JOSS guidelines so the review is not contingent on addressing them:

- You might link your pkg site, https://ctu-bern.github.io/presize from GitHub (and possibly other places)
- There's a typo at the bottom of the Example section in the README, "THe" instead of "The"
- The image at the bottom of your pkgdown site under the Funding header isn't working for me
- I noticed some scenarios where passing nonsense output produces cryptic errors both when calling functions directly or calling them indirectly using the Shiny app. For example, if I mistakenly specified the `conf.width` argument in percent points (e.g., 5) rather than as a proportion (e.g., 0.05),

  ```
  prec_prop(p = 0.5, conf.width = 5, conf.level = 0.95, method = 'wilson')
  Error in uniroot(function(n) eval(quo) - prec, c(1, 1e+07), ...) :
    f() values at end points not of opposite sign
  ```

  or in the Shiny app:

  > **Error:** An error has occurred. Check your logs or contact the app author for clarification.

  I don't think these are deal breakers and I know writing tons of validation logic can be time-consuming but I thought I'd point out that I ran into it.

Please let me know if you have any questions of if I can clarify anything.
