### 2021-03-23 CTU-Bern/presize#61 @TomKellyGenetics
Just a quick note here (I'll submit my review later): Shiny version 1.6.0 appears to be [available on CRAN](https://CRAN.R-project.org/package=shiny) and this issue has been [fixed in version 1.4.0](https://github.com/rstudio/shiny/blob/master/NEWS.md#shiny-140). Is this an issue with the package itself or missing dependencies?
