### 2021-03-23 CTU-Bern/presize#61 @TomKellyGenetics
Thanks for responding so quickly @aghaynes. It makes sense to incorporate changes to address reviews before submitting a new release to CRAN and incrementing the version number associated with the JOSS manuscript to reflect this. On closer reading, this seems to be what @amoeba was suggesting (rather than referring to Shiny versions) as the dev version already solves the missing default values.

I'll test the shiny app (which is a great idea to include!) more and may submit a separate issue if my locale issues persist.
