### 2021-03-26 ba6c131
check prec_prop conf_width (lower only)
