### 2021-03-26 a864bd9
white slider bar ticks/text
