### 2021-03-26 1f0a5ee
remove much of the usage part from paper, rename discussion summary
