+++ paper/paper.md
@@ -27,7 +27,6 @@ date: 2021-02-11
 ---
 
 
-
 # Statement of need
 
 Sample size calculation is a crucial step in planning a clinical study. A too
@@ -47,8 +46,8 @@ We therefore developed a comprehensive tool for precision-based sample size calc
 
 We programmed a package in the R programming language that offers sample size
 calculation for estimation-based research. We implemented the most common
-measures used in descriptive research; see Table 1 for a list of functions,
-corresponding measures, and methods used for calculation. 
+measures used in descriptive research, including descriptive, absolute and relative 
+difference, correlation and diagnostic measures. 
 There are two approaches for each measure. First, based on a given sample size,
 e.g. for a retrospective data analysis, the precision of an expected measure can
 be calculated. Precision is expressed as the confidence interval around the
@@ -61,36 +60,6 @@ further specific input arguments are required such as the expected area under
 the curve (AUC) for test accuracy.
 
 
-Table 1. Measures, functions and methods provided in presize.
-
-Measure | Function | Methods available 
--------- | ---------- | --------
-**Descriptive measures** | |
-Mean | `prec_mean` |
-Proportion | `prec_prop` | Wilson, Agresti-Coull, exact, Wald [see @brown2001]
-Rate | `prec_rate` | Score, variance stabilizing, exact, Wald [see @barker2002]
-**Absolute differences** | |
-Mean difference | `prec_meandiff` |
-Risk difference | `prec_riskdiff` | Newcombe [@newcombe1998], Miettinen-Nurminen [@mn1985], Agresti-Caffo [@ac2000], Wald
-**Relative differences** | |
-Odds ratio | `prec_or` | Gart, Wolff, independence smoothed logit [see @fll2015]
-Risk ratio | `prec_riskratio` | Koopman [@koopman1984], Katz [@kbap1978]
-Rate ratio | `prec_rateratio` | Rothman [@rg2018]
-**Correlation measures** | |
-Correlation coefficient | `prec_cor` | Pearson, Kendall, Spearman [see @bw2000]
-Intraclass correlation | `prec_icc` | @bonnett2002
-Limit of agreement | `prec_lim_agree` | @ba1986
-Cohen's kappa | `prec_kappa` | @rd2012
-**Diagnostic measures** | |
-Sensitivity | `prec_sens` | As per `prec_prop`
-Specificity | `prec_spec` | As per `prec_prop`
-Area under the curve | `prec_auc` | @hm1982
-Negative likelilood ratio | `preg_neg_lr` | @simel1991
-Pogitive likelilood ratio | `preg_pos_lr` | @simel1991
-Generic likelilood ratio | `preg_lr` | @simel1991
-
-
-
 For ease-of-use, we also implemented a Shiny application which can be used
 [online](https://ctu-bern.shinyapps.io/presize) or from within the R-environment.
 Values of parameters can be entered using rulers and numeric fields. Based on
@@ -98,8 +67,6 @@ the given parameters, the application will either display the sample size for a
 given precision, or vice-versa, the precision for a given sample size. Moreover,
 the application also displays the corresponding R-code, which allows copying the
 command into the R-environment for further exploration as well as reproducibility.
-
-
 
 
 # Usage
@@ -112,7 +79,7 @@ command into the R-environment for further exploration as well as reproducibilit
 # remotes::install_github('ctu-bern/presize') # development version on GitHub
 library(presize)
 ```
-Suppose we want to estimate the proportion of hospital admissions with diabetes. 
+As a brief example, suppose we want to estimate the proportion of hospital admissions with diabetes. 
 Diabetes has a prevalence of approximately 10% (@diab). We assume a 
 slightly higher proportion of diabetics, 
[…]
