### 2021-03-26 3e2e4fa
move examples and table from the paper here
