+++ README.Rmd
@@ -53,56 +53,92 @@ Sys.setenv(R_REMOTES_NO_ERRORS_FROM_WARNINGS = "true")
 
 
 ## Overview
-presize will provide functions for
+presize provides functions for
 
-* descriptive statistics
-    + mean (`prec_mean`)
-    + proportion (`prec_prop`)
-    + rate (`prec_rate`)
-  
-* absolute and relative differences
-    + mean difference (`prec_meandiff`)
-    + risk difference (`prec_riskdiff`)
-    + odds ration (`prec_or`)
-    + risk ratio (`prec_riskratio`)
-    + rate ratio (`prec_rateratio`)
-<!--    + hazard ratio -->
-  
-* correlation measures
-    + correlation coefficient (`prec_cor`)
-    + Cohens kappa (`prec_kappa`)
-    + ICC (`prec_icc`)
-    + limit of agreement from Bland Altman plot (`prec_lim_agree`)
-  
-* diagnostic measures
-    + sens (`prec_sens`<sup>1</sup>)
-    + spec (`prec_spec`<sup>1</sup>)
-    + likelihood ratios (`prec_lr`)
-      + positive likelihood ratio (`prec_pos_lr`<sup>2</sup>)
-      + negative likelihood ratio (`prec_neg_lr`<sup>2</sup>)
-    + AUC (`prec_auc`)
-    
+Measure | Function | Methods available 
+-------- | ---------- | --------
+**Descriptive measures** | |
+Mean | `prec_mean` |
+Proportion | `prec_prop` | Wilson, Agresti-Coull, exact, Wald [see @brown2001]
+Rate | `prec_rate` | Score, variance stabilizing, exact, Wald [see @barker2002]
+**Absolute differences** | |
+Mean difference | `prec_meandiff` |
+Risk difference | `prec_riskdiff` | Newcombe [@newcombe1998], Miettinen-Nurminen [@mn1985], Agresti-Caffo [@ac2000], Wald
+**Relative differences** | |
+Odds ratio | `prec_or` | Gart, Wolff, independence smoothed logit [see @fll2015]
+Risk ratio | `prec_riskratio` | Koopman [@koopman1984], Katz [@kbap1978]
+Rate ratio | `prec_rateratio` | Rothman [@rg2018]
+**Correlation measures** | |
+Correlation coefficient | `prec_cor` | Pearson, Kendall, Spearman [see @bw2000]
+Intraclass correlation | `prec_icc` | @bonnett2002
+Limit of agreement | `prec_lim_agree` | @ba1986
+Cohen's kappa | `prec_kappa` | @rd2012
+**Diagnostic measures** | |
+Sensitivity<sup>1</sup> | `prec_sens` | As per `prec_prop`
+Specificity<sup>1</sup> | `prec_spec` | As per `prec_prop`
+Area under the curve | `prec_auc` | @hm1982
+Negative likelilood ratio<sup>2</sup> | `preg_neg_lr` | @simel1991
+Pogitive likelilood ratio<sup>2</sup> | `preg_pos_lr` | @simel1991
+Generic likelilood ratio | `preg_lr` | @simel1991
+
 <sup>1</sup> Simple wrappers for `prec_prop`.
 
 <sup>2</sup> Wrappers for `prec_lr` with values provided via sens and spec
 
 ## Example
 
-This is a basic example which shows you how to solve a common problem:
+Suppose we want to estimate the proportion of hospital admissions with diabetes. 
+Diabetes has a prevalence of approximately 10% (@diab). We assume a 
+slightly higher proportion of diabetics, 
+15%, as diabetes is a risk factor for a wide range of conditions. We want to 
[…]
