+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: presize
 Type: Package
 Title: Precision Based Sample Size Calculation
-Version: 0.2.1
+Version: 0.2.2
 Authors@R: 
     c(person(given = "Armando",
     family = "Lenz",
@@ -27,7 +27,7 @@ Description: Bland (2009) <doi:10.1136/bmj.b3985> recommended to
     functions will return the precision (width of the confidence interval), and 
     vice versa. 
 License: GPL-3 + file LICENSE
-URL: https://github.com/CTU-Bern/presize
+URL: https://github.com/CTU-Bern/presize; https://ctu-bern.github.io/presize
 BugReports: https://github.com/CTU-Bern/presize/issues
 Encoding: UTF-8
 LazyData: true
