### 2021-03-26 dcee5e4
update app link, add references to pkgdown site
