+++ inst/shinyApp/ui/00_helppage.R
@@ -27,16 +27,20 @@ helppage <- tabItem(
   "Relevant references are listed on each page.",
   tags$br(),
   tags$br(),
-  "This site uses the R ",
+  "This site uses the ",
   tags$code("presize"),
   " R package",
-  " (version ",
-  textOutput("version", inline = TRUE),
+  " (version", textOutput("version", inline = TRUE),
   "), which was developed at CTU Bern, the Clinical Trials Unit of the University of Bern and University Hospital Bern, on behalf of the Statistics & Methodology Platform of the ",
   tags$a(href = "https://www.scto.ch/de/network/scto-platforms/statistics-and-methodology.html", "Swiss Clinical Trial Organisation"),
   ". The R package version of ",
   tags$code("presize"),
   " can be installed in R using the from CRAN (",
   tags$code("install.packages('presize')"),
-  "). The R code for running the calculations in this site is shown after the results."
+  "). The R code for running the calculations in this site is shown after the results. The",
+  tags$code("presize"),
+  "package website can be found ",
+  tags$a(href = "https://ctu-bern.github.io/presize/", "here"),
+  "."
+
 )
