### 2021-03-26 1ebc857
upper limit for prec_prop ci_width
