### 2021-03-26 7c708c5
mean sd > 0, prec_rate .1 < ci_width < 5r
