+++ R/shinyApp.R
@@ -5,7 +5,6 @@
 #' @usage
 #' launch_presize_app()
 #' @export
-#' @import shiny ggplot2 shinydashboard
 
 launch_presize_app <- function(){
   shiny::runApp(system.file("shinyApp", package = "presize"))
