### 2021-03-26 1744374
add range checks
