### 2021-03-29 c652e77
input range checks
