### 2021-03-30 779bba7
check boundaries n1 and r
