### 2021-03-30 f06132f
mean diff completely check
