### 2021-03-30 bf4e565
All diff functions done
