+++ R/differences.R
@@ -46,8 +46,10 @@ prec_meandiff <- function(delta, sd1, sd2 = sd1, n1 = NULL, r = 1,
     stop("'r' must be numeric")
   if (sum(sapply(list(n1, conf.width), is.null)) != 1)
     stop("exactly one of 'n', and 'conf.width' must be NULL")
+  
   numrange_check(conf.level)
   numrange_check_gt(r, 0)
+  
   if (!is.null(n1)) {
     numrange_check_gt(n1, 1)
     numrange_check_gt(n1*r, 1)
@@ -58,7 +60,7 @@ prec_meandiff <- function(delta, sd1, sd2 = sd1, n1 = NULL, r = 1,
   if (!is.null(s2)) {
     numrange_check_gt(s2, 0)
   }
-  if (!is.null(conf.width)){
+  if (!is.null(conf.width)) {
     numrange_check_gt(conf.width,0)
   }
   alpha <- 1 - conf.level
@@ -229,13 +231,14 @@ prec_riskdiff <- function(p1, p2, n1 = NULL, conf.width = NULL,
     stop("'p1' must be numeric in [0, 1]")
   if (!is.null(p2) && !is.numeric(p2) || any(0 > p2 | p2 > 1))
     stop("'p2' must be numeric in [0, 1]")
-  if (!is.null(n1)){
+  
+  if (!is.null(n1)) {
     numrange_check_gt(n1,0)
    }
-  if (!is.null(r)){
+  if (!is.null(r)) {
     numrange_check_gt(r,0)
    }
-    if (!is.null(conf.width)){
+    if (!is.null(conf.width)) {
     numrange_check_gt(conf.width,0)
    }
   if (length(method) > 1) {
@@ -510,6 +513,16 @@ prec_riskratio <- function(p1, p2, n1 = NULL, r = 1, conf.width = NULL,
   if (!is.null(p2) && !is.numeric(p2) || any(0 > p2 | p2 > 1))
     stop("'p2' must be numeric in [0, 1]")
 
+  if (!is.null(n1)) {
+    numrange_check_gt(n1,0)
+   }
+  if (!is.null(r)) {
+    numrange_check_gt(r,0)
+   }
+  if (!is.null(conf.width)) {
+    numrange_check_gt(conf.width,0)
+   }
+  
   default_meth <- "koopman"
   if (length(method) > 1) {
     warning("more than one method was chosen, '", default_meth, "' will be used")
@@ -693,7 +706,17 @@ prec_or <- function(p1, p2, n1 = NULL, r = 1, conf.width = NULL, conf.level = 0.
     warning("more than one method was chosen, '", default_meth, "' will be used")
     method <- default_meth
   }
-
+  
+  if (!is.null(n1)) {
+    numrange_check_gt(n1,0)
+  }
+  if (!is.null(r)) {
+    numrange_check_gt(r,0)
+  }
+  if (!is.null(conf.width)) {
+    numrange_check_gt(conf.width,0)
+  }
+  
   meths <- c("gart", "woolf", "indip_smooth")
   id <- pmatch(method, meths)
   meth <- meths[id]
@@ -833,7 +856,22 @@ prec_rateratio <- function(n1 = NULL, # n exposed
     stop("both rate_exp and rate_control required")
   if (sum(sapply(list(n1, prec.level), is.null)) != 1)
     stop("exactly one of 'n1', and 'prec.level' must be NULL")
-
+  if (!is.null(n1)) {
+    numrange_check_gt(n1,0)
+  }
+  if (!is.null(r)) {
+    numrange_check_gt(r,0)
+  }
+  if (!is.null(prec.level)) {
[…]
