### 2021-03-30 3b5eb7a
s1 s2 -> sd1 sd2
