### 2021-03-30 90a6112
correct n1*r error message which was confusing.
also change "if () { }" into one line code
