+++ README.md
@@ -11,6 +11,7 @@ Status](https://ci.appveyor.com/api/projects/status/github/CTU-Bern/presize?bran
 [![Actions
 Status](https://github.com/CTU-Bern/presize/workflows/R-CMD-check/badge.svg)](https://github.com/CTU-Bern/presize/actions)
 [![codecov](https://codecov.io/github/CTU-Bern/presize/branch/master/graphs/badge.svg)](https://codecov.io/github/CTU-Bern/presize)
+[![status](https://joss.theoj.org/papers/780db6ae34eb2317ef9b084012cce4de/status.svg)](https://joss.theoj.org/papers/780db6ae34eb2317ef9b084012cce4de)
 
 [Bland (2009)](https://www.bmj.com/content/339/bmj.b3985) recommended to
 base study sizes on the width of the confidence interval rather the
