### 2021-04-01 CTU-Bern/presize#61 @aghaynes
Once again, thanks for reviewing @amoeba. Here's our response your points. The changes are currently in PR #64 and will hopefully be merged today, once the extra validations are done (this is perhaps relevant to @TomKellyGenetics... maybe you want to review that version?).

## contribution and authorship

 - paper
   - Haynes and Limacher wrote the paper.
   - Stalder put a not-insignificant amount of work into presize testing and we feel that she is thus entitled to coauthorship.
   - Stalder and Lenz provided varying degrees of feedback.

 - R package/git commit history
   - The initial idea for the package came from Limacher, who also secured the funding from the SCTO. Without him, the package probably wouldnt exist.
   - Lenz was the lead package developer until he left CTU ca two years ago.
   - Haynes then took up the lead developer position.
   - Limacher has provided code snippets directly to Lenz or Haynes (sometimes in R, sometimes in STATA), but not written code specifically for the package.
     Limacher also tested functionality throughout the development process.
   - Stalder was a great help in providing support in testing the package against other known approaches but has not authored any of the functions (hence the ctb flag)

 - pkgdown
   - pkgdown defines authors based on info in the DESCRIPTION file.
   - pkgdown only displays those indicated as authors (with the au flag) at the foot of the page, hence only Lenz, Haynes and Limacher are shown.
   - There is a developers section on the right hand side of the page. Clicking this takes you to the "authors" page which details all contributors (au, ctb, any other flags from the DESCRIPTION file).
   - Thus, although not immediately visible, Stalder is given credit on the pkgdown site too.

Although Limacher has not authored code in the package (and thus does not appear in the git commit history), he has been instrumental in its development.

As far as we are concerned, the author lists are correct and appropriate.

## functionality

As already discussed, I found the issue with Shiny recently and fixed it. As far as I can tell, this was due to change in the defaults for the functions in the `shiny` package.

I absolutely agree, the version of the package mentioned in JOSS should have this issue fixed (which it has been in version 0.2.1). Once any other issues are resolved, I will make a release, submit a new version to CRAN and to e.g. zenodo for archiving, as per JOSS requirements.

@majensen, do I need to do anything to tell whedon to use a different version or does it focus on the head version?

## community guidelines

Thanks for pointing this out. I have added sections "Getting help" and "Contributing" to the readme (which are then percolated through to the pkgdown site automatically).

## paper

We added a short summary section at the beginning of the paper.

We have removed most of the "usage" section and the table detailing the different methods to the readme file.
We left a small example in the paper, however.

## non-review details

link to the pkgdown site
- this is a great idea! We have added links to both the readme and the shinyapp.

typo (THe)
- thanks! fixed.

image missing on the pkgdown site
- this should be fixed now.

nonsense input results in cryptic errors
- thanks for pointing this out. We already had a fair amount of validation and are adding more, to basically every function,  which should robustify the app/package against that particular issue. We're still looking at this though.
