### 2021-04-01 #3118 @TomKellyGenetics
Great job on developing a handy package and Shiny App. I think this is deserving of publication with some corrections. The software seems to be of high quality and valuable to the community. The manuscript has some room for improvement. Sorry for the delayed review, I should have time to review in more details soon. I will post minor comments here and major issues on the package repository.

Installation from CRAN works for version 0.1.3 as expected. Installation from GitHub installs the development version 0.2.1 but the documentation is not available. I will review the development version on GitHub conditional on an updated release on CRAN to pending changes during review (as discussed in https://github.com/CTU-Bern/presize/issues/61).
