### 2021-04-01 571deea
Error messages for conf-width
