### 2021-04-01 5720c9e
remove appveyor and travis badges
