+++ README.Rmd
@@ -18,8 +18,6 @@ knitr::opts_chunk$set(
 
 [![](https://www.r-pkg.org/badges/version/presize?color=green)](https://cran.r-project.org/package=presize) 
 `r badger::badge_custom("dev version", as.character(packageVersion("presize")), "blue", "https://github.com/CTU-Bern/presize")`
-![travis](https://travis-ci.com/CTU-Bern/presize.svg?branch=master)
-[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/CTU-Bern/presize?branch=master&svg=true)](https://ci.appveyor.com/project/CTU-Bern/presize)
 [![Actions Status](https://github.com/CTU-Bern/presize/workflows/R-CMD-check/badge.svg)](https://github.com/CTU-Bern/presize/actions)
 [![codecov](https://codecov.io/github/CTU-Bern/presize/branch/master/graphs/badge.svg)](https://codecov.io/github/CTU-Bern/presize)
 [![status](https://joss.theoj.org/papers/780db6ae34eb2317ef9b084012cce4de/status.svg)](https://joss.theoj.org/papers/780db6ae34eb2317ef9b084012cce4de)
