### 2021-04-01 #3118 @aghaynes
Thanks @TomKellyGenetics. As discussed elsewhere, I do indeed intend to submit the newest version to CRAN.
I hope to merge the PR today, pending some modifications for input validation. Note that we made substantial changes to the paper itself in the PR also.
