### 2021-04-01 CTU-Bern/presize#66 @TomKellyGenetics
Improve documentation
## Review: documentation

_Part of [review for JOSS](https://github.com/openjournals/joss-reviews/issues/3118)._

### Examples

The following functions do not have an `@examples` or `\examples` field in their "man" or R files. Please add these so they are displayed in the R/RStudio help system.

- [x] `prec_cor`
- [x] `prec_lim_agree`
- [x] `prec_or`

Ideally several examples should be given, e.g., for Pearson and Spearman correlation and all examples should be well commented to explain the context for understanding use cases.

- [x]  all examples commented

~- [ ]  usage for parameters without defaults explained in examples~

### Shiny App

The documentation for the shiny app needs more details. It is not clear before launching what the app does.

- [x] launch_presize_app

### Building dev version of documentation

Documentation is not available for the dev version 0.2.1 installed from GitHub. These comments above are based on CRAN release 0.1.3 because of this, although the GitHub source seems to have missing examples also. Building the docs with `devtools` or `roxygen2` before pushing to GitHub is recommended (it is required for uploading to CRAN)

- [x] build updated documentation
- [x] check docs are available in R help system before pushing to GitHub

### Suggestions

The review does not hinge on these but are recommended for benefit of users understanding the software.

- [x] **README**:  Providing a screenshot of the Shiny App would give a good preview on GitHub. Images can also [be displayed in the R help system](https://www.r-bloggers.com/2015/12/embedding-images-in-r-package-help-pages/).
- [x] The package itself does not have a documentation page that can be accessed by `help(`presize-package`)` or just `help(precize)`. See the examples "help(`igraph-package`)" or help(`ggplot2-package`).

Optional:

- **Vignette**: A vignette is not required by CRAN but can be helpful to provide further context for examples.
