### 2021-04-01 #3118 @TomKellyGenetics
@aghaynes thanks for letting us know. I will wait for the PR to be merged before checking the updated Shiny App and manuscript. I see you are working on a Docker image as well. Excellent idea. 👍

Regarding the "statement of need", as previously discussed with @majensen and senior editors this does not need to be an explicit section in the paper. Authors may merge this into the abstract/summary at their discretion. The statement of need is clear in the documentation (README).
