+++ R/differences.R
@@ -52,7 +52,7 @@ prec_meandiff <- function(delta, sd1, sd2 = sd1, n1 = NULL, r = 1,
   
   if (!is.null(n1)) numrange_check_gt(n1, 1)
     
-  if (n1*r<=1)
+  if (!is.null(n1) && n1*r<=1)
   stop("'n1'*'r' must be greater than 1")
   
   if (!is.null(sd1)) numrange_check_gt(sd1, 0)
@@ -101,8 +101,15 @@ prec_meandiff <- function(delta, sd1, sd2 = sd1, n1 = NULL, r = 1,
       prec <- eval(md)
     if (is.null(n1)) {
       eqn <- function(r, sd1, sd2, alpha, prec) uniroot(function(n1) eval(md) - prec,
-                                                      c(2, 1e+07), ...)$root
-      n1 <- mapply(eqn, r = r, sd1 = sd1, sd2 = sd2, alpha = alpha, prec = prec)
+                                                      c(2, 1e+07), ...)$root      
+      if(conf.width<min(sd1)){
+        n1 <- try(mapply(eqn, r = r, sd1 = sd1, sd2 = sd2, alpha = alpha, prec = prec), silent = TRUE)
+        if(inherits(n1,"try-error")) stop("'conf.width' too small")
+      } else {
+        n1 <- try(mapply(eqn, r = r, sd1 = sd1, sd2 = sd2, alpha = alpha, prec = prec), silent = TRUE)
+        if(inherits(n1,"try-error")) stop("'conf.width' too wide")
+      }
+      
     }
   }
 
@@ -120,7 +127,13 @@ prec_meandiff <- function(delta, sd1, sd2 = sd1, n1 = NULL, r = 1,
     if (is.null(n1)){
       un <- function(sd1, sd2, r, alpha, prec) uniroot(function(n1) eval(md_ueq) - prec,
                                                       c(2, 1e+07), ...)$root
-      n1 <- mapply(un, sd1 = sd1, sd2 = sd2, r = r, alpha = alpha, prec = prec)
+      if(conf.width<min(sd1)){
+        n1 <- try(mapply(un, sd1 = sd1, sd2 = sd2, r = r, alpha = alpha, prec = prec), silent = TRUE)
+        if(inherits(n1,"try-error")) stop("'conf.width' too small")
+      } else {
+        n1 <- try(mapply(un, sd1 = sd1, sd2 = sd2, r = r, alpha = alpha, prec = prec), silent = TRUE)
+        if(inherits(n1,"try-error")) stop("'conf.width' too wide")
+      }
     }
   }
 
@@ -283,9 +296,17 @@ prec_riskdiff <- function(p1, p2, n1 = NULL, conf.width = NULL,
     if (is.null(n1)) {
       nn <- function(p1, p2, conf.level, conf.width) uniroot(function(n1) eval(nc)$cw - conf.width,
                                                             c(1, 1e+07), ...)$root
-      n1 <- mapply(nn, p1 = p1, p2 = p2, conf.level = conf.level, conf.width = conf.width)
-      ci <- eval(nc)
-      n2 <- ci$n2
+      if(conf.width<min(p1)){
+        n1 <- try(mapply(nn, p1 = p1, p2 = p2, conf.level = conf.level, conf.width = conf.width), silent = TRUE)
+        if(inherits(n1,"try-error")) stop("'conf.width' too small")
+        ci <- eval(nc)
+        n2 <- ci$n2
+      } else {
+        n1 <- try(mapply(nn, p1 = p1, p2 = p2, conf.level = conf.level, conf.width = conf.width), silent = TRUE)
[…]
