### 2021-04-02 CTU-Bern/presize#67 @TomKellyGenetics
Improve community guidelines
## Review: Community guidelines

_Part of [review for JOSS](https://github.com/openjournals/joss-reviews/issues/3118)._

### Contributing guidelines

Regarding this criteria, the [Contributing guidleines](https://github.com/CTU-Bern/presize/blob/master/CONTRIBUTING.md) are available (as requested in https://github.com/CTU-Bern/presize/issues/61).
 > 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

However, efforts should be made to direct users to these and guide new users through this process (I note some have been made already).

- [x] provide contributing guidelines
- [x] direct users to them in the README
- [x]  direct users to the GitHub repository from CRAN (DESCRIPTION)
- [x] direct users to the Issues for Bug Reports from CRAN (DESCRIPTION)
- [x] GitHub issues template for bug reports and feature requests
- [x] GitHub pull request template

Some of these are not strictly necessary for publication but are minor efforts that aid novice R users for bug reporting or seeking support via GitHub. See the [community](https://github.com/CTU-Bern/presize/community) GitHub page for further recommendations.
