### 2021-04-06 b18c395
Comment mean_diff check
