### 2021-04-06 CTU-Bern/presize#61 @amoeba
Sorry for the delayed response @aghaynes. Thanks for taking the time to work over my review and make the changes. Extra thanks for making those changes in a PR so I can check them out easily.

- Re: paper/authorship: That sounds great to me. Thanks for laying that out in so much detail and thanks for confirming things are correct.
- functionality: 👍
- community guidelines: 👍
- non-review details: 👍
- paper: Thanks, those changes improve things. The paper reads very well, is nicely-written, and I'd sign on off it except for this section from the [Review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html):

  > Note the paper should not include software documentation such as API (Application Programming Interface) functionality, as this should be outlined in the software documentation.

  I think the Usage section violates the intent of this rule. Instead of guessing, maybe we could pull in @majensen.

The only outstanding part of my review is the above question about the paper. Let's see what @majensen says and go from there?
