### 2021-04-06 CTU-Bern/presize#61 @TomKellyGenetics
Regarding this point, @majensen previously edited [this paper](https://joss.theoj.org/papers/10.21105/joss.02161) of mine which also includes R code. I agree that the examples in the documentation should stand on their own but usage sections in R papers often follow a "vignette" style like this.

PR #64 seems to pass CI checks so if it is merged I'm available to review the updated manuscript.
