### 2021-04-06 5644fad
update help, add figure to app helppage
