+++ man/presize-package.Rd
@@ -16,7 +16,8 @@ Bland (2009) <doi:10.1136/bmj.b3985> recommended to
 \seealso{
 Useful links:
 \itemize{
-  \item \url{https://github.com/CTU-Bern/presize; https://ctu-bern.github.io/presize}
+  \item \url{https://github.com/CTU-Bern/presize}
+  \item \url{https://ctu-bern.github.io/presize}
   \item Report bugs at \url{https://github.com/CTU-Bern/presize/issues}
 }
 
