+++ DESCRIPTION
@@ -27,7 +27,7 @@ Description: Bland (2009) <doi:10.1136/bmj.b3985> recommended to
     functions will return the precision (width of the confidence interval), and 
     vice versa. 
 License: GPL-3 + file LICENSE
-URL: https://github.com/CTU-Bern/presize; https://ctu-bern.github.io/presize
+URL: https://github.com/CTU-Bern/presize, https://ctu-bern.github.io/presize
 BugReports: https://github.com/CTU-Bern/presize/issues
 Encoding: UTF-8
 LazyData: true
@@ -35,7 +35,8 @@ RoxygenNote: 7.1.1
 Suggests:
     binom,
     Hmisc,
-    testthat
+    testthat,
+    shinytest
 Imports:
     ggplot2,
     kappaSize (>= 1.2),
