+++ man/prec_or.Rd
@@ -61,7 +61,7 @@ recommended intervals (Fagerland et al. 2011).
 indip_smooth method.
 }
 \examples{
-# 10\% events in one group, 15\% in the other, 200 participants total
+# 10\\% events in one group, 15\\% in the other, 200 participants total
 #  (= 100 in each group), estimate confidence interval width
 prec_or(p1 = .1, p2 = .15, n1 = 200/2)
 # formula by Gart
@@ -69,7 +69,7 @@ prec_or(p1 = .1, p2 = .15, n1 = 200/2, method = "gart")
 # formula by Woolf
 prec_or(p1 = .1, p2 = .15, n1 = 200/2, method = "woolf")
 
-# 10\% odds in one group, 15\% in the other, desired CI width of 0.1,
+# 10\\% odds in one group, 15\\% in the other, desired CI width of 0.1,
    estimate N
 prec_or(p1 = .1, p2 = .15, conf.width = .1)
 # formula by Gart
