+++ man/prec_cor.Rd
@@ -52,18 +52,18 @@ n is rounded up to the next whole number using \code{ceiling}.
 \examples{
 # calculate confidence interval width...
 # Pearson correlation coefficient
-prec_cor(.5, n = 100)
+prec_cor(r = 0.5, n = 100)
 # Kendall rank correlation coefficient (tau)
-prec_cor(.5, n = 100, method = "kendall")
+prec_cor(r = 0.5, n = 100, method = "kendall")
 # Spearman's rank correlation coefficient
-prec_cor(.5, n = 100, method = "spearman")
+prec_cor(r = 0.5, n = 100, method = "spearman")
 # calculate N required for a given confidence interval width...
 # Pearson correlation coefficient
-prec_cor(conf.width = .15, n = 100)
+prec_cor(r = 0.5, conf.width = .15)
 # Kendall rank correlation coefficient (tau)
-prec_cor(conf.width = .15, n = 100, method = "kendall")
+prec_cor(r = 0.5, conf.width = .15, method = "kendall")
 # Spearman's rank correlation coefficient
-prec_cor(conf.width = .15, n = 100, method = "spearman")
+prec_cor(r = 0.5, conf.width = .15, method = "spearman")
 }
 \references{
 Bonett DG, and Wright TA (2000) \emph{Sample size requirements
