+++ CONTRIBUTING.md
@@ -1,34 +1,0 @@
-# Contributing to presize
-
-This guide gives some points to contributing to `presize`. Contributions do not
-have to entail literally modifying/adding code, bug reports are equally
-important.
-
-* Bug reports should be made via issues
-
-* Modifications or additions to code can be made via pull requests (PR)
-
-## Issues
-
-Be concise, but explain what the problem is. Adding code and output is useful,
-if possible.
-Consider commenting the code to improve clarity.
-
-
-## Pull requests
-
-To contribute modifications or additions to code, create a PR. The
-steps are broadly as follows.
-
-1. Fork `presize` to your GitHub profile.
-1. Clone it to your local machine.
-1. Make your changes and commit them to your local git.
-1. Push your changes to GitHub.
-1. Create a PR
-1. Discuss with reviewer about any requested changes until the PR can be
-closed.
-
-Please describe the motivation behind your PR and provide unit tests via
-testthat. If your PR addresses an issue, add that info in the description
-e.g. closes #{issue-number}
-
