### 2021-04-06 CTU-Bern/presize#66 @aghaynes
Thanks @TomKellyGenetics.

I have added examples for the three functions you mentioned and added extra info and the screenshot to the launch_shiny_app help file. I've also added a presize-package helpfile and the app screenshot to the readme (as per your suggestions).

A quick question regarding the examples... you say "usage for parameters without defaults explained in examples"... as the examples are on the help pages and the help pages already explain the parameters, isn't this a bit over the top? Examples in most helpfiles are very minimal...

I'm actually quite surprised that the helpfiles aren't there... I can see them when I install through remotes...
In general, one of our github actions actually builds all our package documentation for us (including the pkgdown site) each time something is merged in to the main branch, so everything should be up to date.... although this might not be the case for 0.2.2 on my fork...

Regarding your vignette suggestion, it would indeed be nice, but to do it well would require a little bit of time I think. Something to think about for the (near) future... is that OK?
