### 2021-04-06 CTU-Bern/presize#67 @aghaynes
Thanks @TomKellyGenetics.

I've added issue templates (already in the main fork) and a PR template (in PR #61) and added a link on the readme to the contributing md file (also in PR #61).

Thanks also for pointing out the community page... I'd not noticed that...
