### 2021-04-06 CTU-Bern/presize#67 @aghaynes
@TomKellyGenetics, I've merged #61. Documentation typically takes about 10 mins to compile.
