### 2021-04-06 CTU-Bern/presize#61 @aghaynes
Hi both, I've merged #61 into the main branch. The shiny app online takes a little more effort to update... thats something for tomorrow or thursday I think.
