### 2021-04-06 CTU-Bern/presize#61 @majensen
Hi @amoeba @TomKellyGenetics @aghaynes  -- my feeling about R is that code in the docs is part of its culture. It's natural to give an example in the paper, it not "documentation" per se. So unless I'm overridden by a higher power, I extend to you the scepter of approval.
