### 2021-04-07 CTU-Bern/presize#66 @TomKellyGenetics
Installing `remotes::install_github("CTU-Bern/presize")` installs presize version 0.2.1 and  `remotes::install_github("aghaynes/presize")` installs versions 0.2.2. In both cases the documentation is now available in R.  👍

Is the `pgkdown` and `devtools::document()` workflow triggered when merging to `CTU-Bern/presize`? I see the examples are updated on your branch with comments in the code. Presuming this is merged into the package, this is acceptable.

I'd prefer more explanation of the examples such as how these apply to real-world data or study design. A vignette would be the best place for long-form documentation like this. I understand it would take more time. As examples are also explained well in the README and manuscript, this is not a strict requirement for publication.
