### 2021-04-07 CTU-Bern/presize#67 @TomKellyGenetics
Thanks, I wasn't aware of the GitHub [Discussions](https://github.com/CTU-Bern/presize/discussions) feature, it seems  useful for engaging with the user community! Some of what you have there to discuss with the developers might be better suited to [Projects](https://github.com/CTU-Bern/presize/projects) though.
