### 2021-04-07 30b5e6e
mask variables for output
