+++ README.Rmd
@@ -78,7 +78,7 @@ Sensitivity<sup>1</sup> | `prec_sens` | As per `prec_prop`
 Specificity<sup>1</sup> | `prec_spec` | As per `prec_prop`
 Area under the curve | `prec_auc` | @hm1982
 Negative likelilood ratio<sup>2</sup> | `preg_neg_lr` | @simel1991
-Pogitive likelilood ratio<sup>2</sup> | `preg_pos_lr` | @simel1991
+Positive likelilood ratio<sup>2</sup> | `preg_pos_lr` | @simel1991
 Generic likelilood ratio | `preg_lr` | @simel1991
 
 <sup>1</sup> Simple wrappers for `prec_prop`.
