### 2021-04-07 CTU-Bern/presize#71 @TomKellyGenetics
Improve manuscript for JOSS Review
## Review: manuscript

_Part of [review for JOSS](https://github.com/openjournals/joss-reviews/issues/3118)._

### Overall

Firstly, let me thank the authors for their efforts to prepare an excellent package. It's certainly of wide interest for the research community and particularly providing a Shiny App is a great way to make it accessible to those without programming expertise. As R is popular in statistics, life sciences and biomedical research, an R package would be widely used in these fields.

I agree that PR #64 significantly improves the manuscript as discussion in Issue #61. Overall the paper nicely summarises the functionality of the package and provides some examples of use cases. I have some minor feedback on structure and grammar. Generally, I believe the paper is worth of acceptance with minor editing.

### Contributions

> - [x] Contribution and authorship: Has the submitting author (@aghaynes) made major contributions to the software Does the full list of paper authors seem appropriate and complete?
> - [x] Substantial scholarly effort: Does this submission meet the scope eligibility described in the JOSS guidelines

As shown [here](https://github.com/CTU-Bern/presize/pull/56) @aghaynes has prepared the manuscript and got feedback other authors. Authorship seems to be appropriate to me.

### Structure

> - [x] Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
> - [x] A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?

These are both well covered but I'd recommend merging them into one "Summary" section  to avoid repetition. For example [lines 9, 19, and 22](https://github.com/CTU-Bern/presize/blame/master/paper/paper.md#L32-L55) are similar.

## Introduction

> - [x] State of the field: Do the authors describe how this software compares to other commonly-used packages?

Normally I'd expect some discussion of alternatives but none seem to be available for "precision" sample sizes (or confidence intervals). I'd be willing to make an exception to this as under the circumstances but it could be helpful to provide some examples of "power" calculations for hypothesis testing. _JOSS_ manuscripts are meant to be short but I think some comparisons here would be informative and related software (e.g.,  for power calculations in R) should be cited. As pointed out the paper this is complementary work.

## Usage

As discussed in #61 I think this section is appropriate, especially as long-form vignette documentation (discussed here: #66) is not yet available.

## Grammar

The writing is overall of high quality and easy to understand for non-experts. Some statistical jargon is unavoidable in this kind of work. There are a few sections that could be rephrased for clarity. These should be minor changes that will not take long. Please read the manuscript carefully for similar issues.

> line 9, 20: we **have** therefore...

> line 13: Sample size is a crucial step ~~in~~ **for** planning a clinical study.

> line 13-14: A ~~too small~~ study **that is too small** leads to inconclusive results.

> line 18: man research projects aim ~~at~~ **to**...

> line 22: We ~~programmed~~ **have developed an R package** that...

Change "first" and "second" to "firstly" and "secondly".

> line 35: we **have** also

> line 42: can be **installed and** loaded into the R session...

> line 53: we **have** developed

> ~~line 61: We often observe in our consulting activity that researchers try to implement a hypothesis-based approach~~

> line 64 fascilitate -> facilitate, adequate -> appropriate

- [x] correct spelling and grammar (above)

### References

References are formatted correctly but a few notable citations are missing.

- [x] Cite R packages for power calculations
- [x] Cite the R core team

```
> citation()

To cite R in publications use:

  R Core Team (2020). R: A language and environment for statistical computing. R Foundation for
  Statistical Computing, Vienna, Austria. URL https://www.R-project.org/.

A BibTeX entry for LaTeX users is

  @Manual{,
    title = {R: A Language and Environment for Statistical Computing},
    author = {{R Core Team}},
    organization = {R Foundation for Statistical Computing},
    address = {Vienna, Austria},
    year = {2020},
    url = {https://www.R-project.org/},
  }

We have invested a lot of time and effort in creating R, please cite it when using it for data
analysis. See also ‘citation("pkgname")’ for citing R packages.
```

### Questions

> A limitation of the package is that it does not allow calculating the probability of a CI, i.e. the probability that a future confidence interval will have at least the desired precision.

It's great that the authors have acknowledged clearly and as it stands the packages is already very useful. Are there plans to add this in the future? How difficult would it be? Maintaining an R package is a commitment and I understand that time and funding to support this may be limited. As mentioned in the paper, investing in this will reduce the burden on consulting activities and aid researchers that do not have this expertise available.

-----------------

This completes my review of the manuscript. I will test the latest version of the software and shiny app to complete the _JOSS_ review.
