### 2021-04-07 CTU-Bern/presize#72 @TomKellyGenetics
Reporting error: sensitivity and specificity in shiny app
**Describe the bug**
Shiny app () returns error as shown:

```ntot' required when 'prev' is specified```

**To Reproduce**
See screenshots. Hosted locally with:

```
launch_presize_app()
```

Development version installed with `remotes::install_github("aghaynes/presize")`. Commit: https://github.com/aghaynes/presize/commit/a917c32fc886dfd2f8eee9f7c6743f2803134cf8

**Expected behavior**
Prevalence input is required by ShinyApp. It is possible to use a button to disable it. Slider could be removed from UI with conditional input.

Estimating sample size from `conf.width` without specifying `n` is not currently supported in Shiny App. Cannot specify `ntot` from Shiny App.

**Screenshots**

![Screen Shot 2021-04-07 at 15 05 15](https://user-images.githubusercontent.com/5493325/113818999-a3530800-97b3-11eb-81f7-96e8a5a2d5ee.png)
![Screen Shot 2021-04-07 at 15 05 30](https://user-images.githubusercontent.com/5493325/113819009-a64df880-97b3-11eb-8303-296f9b84efc4.png)

```
R version 4.0.3 (2020-10-10)
 os       macOS Catalina 10.15.7
 system   x86_64, darwin17.0
```

```
   presize        * 0.2.2      2021-04-07 [1] Github (aghaynes/presize@a917c32)
```
