### 2021-04-07 CTU-Bern/presize#72 @TomKellyGenetics
As a note on the Shiny App. The locale issue discussed in #61 seems to be resolved now. The Shiny App launches in various locale settings and I cannot reproduce the error (possibly caused by dependencies).

```
> Sys.setenv(LANG = "fr")
> 2 = x
Error in 2 = x : membre gauche de l'assignation (do_set) incorrect

> presize::launch_presize_app()

Listening on http://127.0.0.1:6747

> Sys.setenv(LANG = "ja")
> 2  + x
 エラー:  オブジェクト 'x' がありません

> presize::launch_presize_app()

Listening on http://127.0.0.1:6747

> Sys.setenv(LANG = "")
> presize::launch_presize_app()

Listening on http://127.0.0.1:6747

> Sys.setenv(LANG = "en")
> 2 + x
Error: object 'x' not found

> presize::launch_presize_app()

Listening on http://127.0.0.1:6747

> Sys.setenv(LANG = "UTF8")
> 2+ x
Error: object 'x' not found
> presize::launch_presize_app()

Listening on http://127.0.0.1:6747
```
