### 2021-04-07 #3118 @TomKellyGenetics
I've reported a bug here: https://github.com/CTU-Bern/presize/issues/72

This completes my review. I recommend acceptance conditional on revisions in issues opened on the https://github.com/CTU-Bern/presize repository. Great work to develop a useful package.
