### 2021-04-07 CTU-Bern/presize#66 @aghaynes
It is odd that for you, installing via `remotes::install_github("CTU-Bern/presize")` installs v0.2.1... I get 0.2.2, as it should be...

```
> remotes::install_github("ctu-bern/presize")
Downloading GitHub repo ctu-bern/presize@HEAD
[truncated for brevity...]
** testing if installed package keeps a record of temporary installation path
* DONE (presize)
> packageVersion("presize")
[1] ‘0.2.2’
```

Yes, `pkgdown` and `devtools::document` are both triggered automatically on a merge (technically a push, but that's the same thing in a PR). It's handled by [this action](https://github.com/CTU-Bern/presize/blob/master/.github/workflows/render-readme.yaml).

Re the vignette, can we consider the issue resolved if we agree to include a vignette in the next version (0.2.3 presumably, assuming CRAN accept the current version as is) of the package?

One question regarding html vignettes (as I notice that you have a few on at least one of your packages), if I may... in another package, we tried to include a html vignette and got a message back from CRAN that you have to indicate external dependencies (such as pandoc; they even set the package as to-be-archived, so we had to reformat everything for PDF vignettes). I've looked at other packages with html vignettes and could never determine where that is done... googling has also never helped me. Can you shed some light on that for me?
