### 2021-04-07 CTU-Bern/presize#66 @TomKellyGenetics
You're correct, installing from GItHub now gives the correct version:
```
> packageVersion("presize")
[1] '0.2.2'
```

A possible explanation for this is that I cloned the repository and run `devtools` locally to build the documentation. I may have been using an outdated version last week if the package was updated in the meantime.

I'm okay for version 0.2.2 to be published without a vignette, although I would encourage it for the next release. For HTML vignettes you have 2 options.

1. Provide vignettes in Rmarkdown. They will be automatically built as HTML on CRAN the header of the Rmarkdown file will tell CRAN which Vignette builder to use. See [this example](https://github.com/TomKellyGenetics/vioplot/blob/master/vignettes/violin_customisation.Rmd). You can check vignette building with devtools.

2. You can provide pre-generated [static HTML vignettes](https://cran.r-project.org/web/packages/R.rsp/vignettes/R_packages-Static_PDF_and_HTML_vignettes.pdf) with some [modifications](https://github.com/HenrikBengtsson/R.rsp/issues/42) these are CRAN compatible.

The current docs are acceptable to me, although I understand some changes may be needed to pass CRAN checks.
