+++ paper/paper.md
@@ -28,38 +28,36 @@ date: 2021-02-11
 
 # Summary
 
-Sample size calculation based on precision rather than hypothesis testing is 
-underused and respective tools not widely available. We therefore developed a 
-software package, `presize`, for precision-based sample size calculation that 
-covers most common summary measures and can be used with the R-environment or a 
-Shiny application.
-
-# Statement of need
-
-Sample size calculation is a crucial step in planning a clinical study. A too
-small study leads to inconclusive results; a too large study is a waste of
-resources. Either case might be unethical. 
+Sample size calculation is a crucial step for planning a clinical study. A study 
+that is too small leads to inconclusive results; a study that is too large is a 
+waste of resources. Either case might be unethical. 
 Furthermore, @bland2009 called for a focus on the width of confidence intervals 
-rather than the power of the test in sample size calculations. 
-There are many software packages
-for hypothesis-based sample size calculation, such as [Stata](https://www.stata.com/),
-[PASS](https://www.ncss.com/software/pass/), and [G*Power](https://www.gpower.hhu.de).
-However, many research projects aim at estimation rather than hypothesis testing,
+rather than the power of the test in sample size calculations. Indeed, many 
+research projects aim to estimate a quantity rather than test a hypothesis,
 sample size calculation approaches for which are largely missing from other software
-packages. 
-We therefore developed a comprehensive tool for precision-based sample size calculation.
+packages. There are many software packages for hypothesis-based sample size calculation, 
+such as [Stata](https://www.stata.com/), [PASS](https://www.ncss.com/software/pass/),
+[G*Power](https://www.gpower.hhu.de), including many R packages, such as 
+[`pwr`](https://CRAN.R-project.org/package=pwr) (@pwr) and
+[`TrialSize`](https://CRAN.R-project.org/package=TrialSize) 
+(@trialsize; see other packages detailed on the 
+[CRAN Clinical Trials taskview](https://cran.r-project.org/web/views/ClinicalTrials.html)).
+To the best of our knowledge, only Stata provides precision-based approaches, and 
+only then for a small number of statistics.
+We have, therefore, developed an R package for precision-based sample size calculation, 
+`presize`, which can be used within the R-environment or a shiny application.
 
 # Development
 
-We programmed a package in the R programming language that offers sample size
+`presize` is programmed in the R programming language (@cran), and offers sample size
 calculation for estimation-based research. We implemented the most common
 measures used in descriptive research, including descriptive, absolute and relative 
-difference, correlation and diagnostic measures. 
-There are two approaches for each measure. First, based on a given sample size,
+differences, correlation and diagnostic measures. 
+There are two approaches for each measure. Firstly, based on a given sample size,
 e.g. for a retrospective data analysis, the precision of an expected measure can
 be calculated. Precision is expressed as the confidence interval around the
 measure. The level of confidence can be specified; the usual 95%-confidence
-interval (CI) is the default. Second, based on a given precision (i.e. CI), the 
+interval (CI) is the default. Secondly, based on a given precision (i.e. CI), the 
 sample size can be calculated. This is mainly of use in the planning of prospective 
 studies, where the aim is to estimate a measure of interest with enough confidence. 
 The available functions in the package require common input arguments; either the 
@@ -68,7 +66,7 @@ function, further specific input arguments are required such as the expected are
 under the curve (AUC) for test accuracy.
 
 
-For ease-of-use, we also implemented a Shiny application which can be used
+For ease-of-use, we have also implemented a Shiny application which can be used
 [online](https://ctu-bern.shinyapps.io/presize) or from within the R-environment.
[…]
