### 2021-04-07 CTU-Bern/presize#71 @aghaynes
Thanks for the additional comments @TomKellyGenetics. I have incorporated them into #73.

I merged the statement of need and summary and remodelled it slightly to improve the flow. I fixed the other comments as you indicated with the exception of

> line 61: We often observe in our consulting activity that researchers try to implement a
hypothesis-based approach

as I don't see anything wrong there (MS Word also sees no issue with it...).

I highlight the `pwr` and `TrialSize` packages and also the Clinical Trials task view which lists a large number of packages that might be of relevance (although none cover precision based methods to our knowledge). I also cite R - thanks for pointing that out!

Furthermore, I also add you, @amoeba and @majensen to the acknowlegdements.

Regarding your question about the probability of an estimate being within the CI... there is a [paper](https://pubmed.ncbi.nlm.nih.gov/8017747/) describing the method in the appendix if I remember but we need to think hard about how to do it for each of the different methods. It is something that we are considering adding in the future though, when time allows. Stata actually has the option for the few methods available there, although there code is proprietory of course and not (easily) accessible, if at all.
