### 2021-04-07 #3118 @amoeba
Hello @majensen my review is now an Accept. We noticed an issue with the 0.1.3 version that's both on CRAN and the version being archived with this submission so I'd like to see a new version archived if possible once @aghaynes has minted it.
