### 2021-04-08 CTU-Bern/presize#66 @TomKellyGenetics
If you have more questions about the vignettes, please let me know! Otherwise the issues raised here have been addressed.
