### 2021-04-08 #3118 @TomKellyGenetics
Thanks for responding to reviews so promptly @aghaynes, great work! As discussed in https://github.com/CTU-Bern/presize/issues/66 version 0.2.2 will be "published" and released on CRAN. Everything seems to be in order but I'll check the updated manuscript once whedon has updated for PR https://github.com/CTU-Bern/presize/pull/73.
