### 2021-04-08 CTU-Bern/presize#71 @TomKellyGenetics
I've checked the updated paper and everything seems to be taken care of.

As a stylistic suggestion, I think this section would fit better in the Introductory section (following paragraph). It's a bit detailed for the "Summary".

>  There are many software packages for hypothesis-based sample size calculation, such as [Stata](https://www.stata.com/), [PASS](https://www.ncss.com/software/pass/),
[G*Power](https://www.gpower.hhu.de), including many R packages, such as [`pwr`](https://CRAN.R-project.org/package=pwr) (@pwr) and[`TrialSize`](https://CRAN.R-project.org/package=TrialSize) (@trialsize; see other packages detailed on the [CRAN Clinical Trials taskview](https://cran.r-project.org/web/views/ClinicalTrials.html)). To the best of our knowledge, only Stata provides precision-based approaches, and only then for a small number of statistics.

I accept the manuscript as is and leave this change to the author's discretion.
