### 2021-04-08 CTU-Bern/presize#72 @TomKellyGenetics
I've tested installing from GItHub and verified that the Shiny App now works as expected. 👍
