### 2021-04-08 #3118 @TomKellyGenetics
I've installed the development version 0.2.2 of `presize` from GitHub and verified that [updated documentation](https://github.com/CTU-Bern/presize/issues/66) is available and the [issue with the Shiny App](https://github.com/CTU-Bern/presize/issues/72) has been corrected. All [comments](https://github.com/CTU-Bern/presize/issues/71) on the manuscript have been satisfactorily addressed.

@majensen my review is also "accept". Well done @aghaynes and colleagues.
