### 2021-04-08 #3118 @aghaynes
Great, thanks for everyones help! I'll prep the CRAN submission now...

@majensen, note that the version number for the paper should be changed to 0.2.2
