### 2021-04-08 1c68bc6
move ggplot and shinydashboard to suggests
