+++ DESCRIPTION
@@ -34,11 +34,11 @@ LazyData: true
 RoxygenNote: 7.1.1
 Suggests:
     binom,
+    ggplot2,
     Hmisc,
     testthat,
+    shinydashboard,
     shinytest
 Imports:
-    ggplot2,
     kappaSize (>= 1.2),
-    shiny,
-    shinydashboard
+    shiny
