### 2021-04-08 #3118 @TomKellyGenetics
I seem to be a more friendly timezone for you than the editors so I'll let you know so you can get started on the next step of the submission process. Can you create an archive for this version on Zenodo, Figshare, or similar (see https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729 for a tutorial if needed)? Please make sure authors and the title of the archive match those of the paper and give the DOI here. Doing it at the same time as the CRAN release will ensure the same version is used too. 😀
