### 2021-04-08 81249bf
changes to pass CRAN checks
