### 2021-04-08 #3118 @aghaynes
The package is with CRAN now (version is now 0.2.3).

Thanks for the tips @TomKellyGenetics. I'll wait to see if there are any editorial PRs before doing Zenodo.
