### 2021-04-09 #3118 @aghaynes
FYI, the package has been accepted by CRAN (version 0.2.3).
