### 2021-04-10 #3118 @majensen
@aghaynes I have only two minor issues:
- proof line number 15 + 86: Can you fix the reference so that it formats "Wu, V. Q., Chow, S.-C., ... " ?
- proof line number 69: you want the possessive "its", rather than contraction "it's".
After those tweaks, I say let's make the archive. Please report the resulting DOI here. Thanks!
