+++ paper/paper.bib
@@ -201,7 +201,7 @@
 @Manual{trialsize,
   title = {TrialSize: R Functions for Chapter 3,4,6,7,9,10,11,12,14,15 of Sample Size
 Calculation in Clinical Research},
-  author = {Ed Zhang ; Vicky Qian Wu ; Shein-Chung Chow ; Harry G.Zhang},
+  author = {Zhang, Ed and Wu, Vicky Qian and Chow, Shein-Chung and Zhang, Harry G.},
   year = {2020},
   note = {R package version 1.4},
   url = {https://CRAN.R-project.org/package=TrialSize},
