### 2021-04-12 #3118 @kyleniemeyer
Hi @aghaynes, I am the AEIC on duty this week doing some final checks before accepting your paper.

It looks like your article is currently missing a Statement of Need section, a [required element](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) of JOSS papers that " clearly illustrates the research purpose of the software." Usually our bot identifies when these are missing very early on, but I think the initial problems with your paper compiling might have led to it being missed then.

Could you please add this section? You can likely just move some of the text in your Summary section into this new section.

Also, it looks like some of your citations are using the incorrect command; for example, "(R Core Team (2020))" on the first page and "(Emerging Risk Factors Collaboration et al. (2010))" on the second page. For these, rather than manually wrapping in parentheses, use the syntax `[@author:2001]`, which will appear as "(Author et al., 2001)". Could you fix these?
