### 2021-04-13 #3118 @TomKellyGenetics
To clarify @kyleniemeyer, the article did have a "Statement of Need" section in earlier version. Based on [previous discussions with the editor](https://github.com/openjournals/joss-reviews/issues/2161#issuecomment-603312770), I suggested removing the heading and merging into the "Summary" section. Sorry if this has caused confusion.
