### 2021-04-14 #3118 @kyleniemeyer
> @aghaynes I'd say since the code won't change there's no need for a new archive.

Yes, this is right—JOSS archives the paper itself, and so we can refer to the existing Zenodo archive for the software.
