### 2021-04-14 #3118 @aghaynes
OK, i think it's ready again...

There is one reference thats still a little odd (the one for `trialsize`), but i dont know how to fix that. I tried the [@author:year; other stuff] syntax, but that didnt compile properly (it didnt work for any of the references I tried it for).

![image](https://user-images.githubusercontent.com/6594028/114660624-ec481680-9cf5-11eb-82bc-e3026e528eb4.png)
