### 2021-04-14 #3118 @kyleniemeyer
Many thanks to @amoeba and @TomKellyGenetics for reviewing this, and @majensen for editing it.

(FYI, the DOI seems to still not be resolving, but this is likely just a delay. Everything else on the JOSS side looks good.)
