### 2021-04-14 #3118 @kyleniemeyer
Looks like the DOI is still not resolving, which is likely a delay with CrossRef. In the meantime you can see the published article here: https://joss.theoj.org/papers/10.21105/joss.03118
